Extract navbar
