+++ components/page.js
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import styled from 'styled-components';
-import Link from 'next/link';
 import Sidebar from './sidebar';
+import Navbar from './navbar';
 import { COLOR } from '../constants';
 import LoadingGif from '../assets/nwplus.gif';
 
@@ -31,50 +31,14 @@ const Content = styled.div`
   padding: 60px;
 `;
 
-const HackathonNavBar = styled.div`
-  display: flex;
-  padding-left: 60px;
-`;
-
-const NavItem = styled.a`
-  display: block;
-  font-size: 24px;
-  font-weight: 600;
-  margin-right: 40px;
-  text-decoration: none;
-  color: ${(p) => (p.selected ? COLOR.BLACK : COLOR.DARK_COPY)};
-  border-bottom: ${(p) => (p.selected ? `3px solid ${COLOR.BLACK}` : 'none')};
-  transition: color 0.4s cubic-bezier(0.25, 0.8, 0.25, 1);
-  &:hover {
-    color: ${COLOR.BLACK};
-  }
-  &:focus {
-    color: ${COLOR.BLACK};
-  }
-`;
-
 export default ({ hackathons, currentPath, children }) => {
-  const [currPath, setCurrPath] = useState('intro');
   const [loading, setLoading] = useState(false);
   const [timeOut, setTimeOut] = useState();
-  useEffect(() => {
-    setCurrPath(window.location.pathname.split('/')[2]);
-  }, []);
 
   useEffect(() => {
     setLoading(false);
     clearTimeout(timeOut);
-  }, [currPath]);
-
-  const onClick = ({ path }) => {
-    if (!window.location.href.includes(path)) {
-      setTimeOut(
-        setTimeout(() => {
-          setLoading(true);
-        }, 750)
-      );
-    }
-  };
+  }, [currentPath]);
 
   return (
     <Container>
@@ -84,44 +48,11 @@ export default ({ hackathons, currentPath, children }) => {
           <Header>{currentPath}</Header>
           {loading && <LoadingImage src={LoadingGif} />}
         </HeaderContainer>
-        <HackathonNavBar>
-          <Link href="/[id]/intro" as={`/${currentPath}/intro`} passHref>
-            <NavItem
-              onClick={() => onClick('intro')}
-              selected={currPath === 'intro'}
-            >
-              Intro
-            </NavItem>
-          </Link>
-          <Link href="/[id]/faq" as={`/${currentPath}/faq`} passHref>
-            <NavItem
-              onClick={() => onClick('faq')}
-              selected={currPath === 'faq'}
-            >
-              FAQ
-            </NavItem>
-          </Link>
-          <Link href="/[id]/spocos" as={`/${currentPath}/spocos`} passHref>
-            <NavItem
-              onClick={() => onClick('spocos')}
-              selected={currPath === 'spocos'}
-            >
-              Sponsors
-            </NavItem>
-          </Link>
-          <Link
-            href="/[id]/FeatureFlags"
-            as={`/${currentPath}/FeatureFlags`}
-            passHref
-          >
-            <NavItem
-              onClick={() => onClick('FeatureFlags')}
-              selected={currPath === 'FeatureFlags'}
-            >
-              FeatureFlags
-            </NavItem>
-          </Link>
-        </HackathonNavBar>
+        <Navbar
+          setLoading={setLoading}
+          currentPath={currentPath}
+          setTimeOut={setTimeOut}
+        />
         <Content>{children}</Content>
       </div>
     </Container>
